SCT relax postal address fields for payments
Update the **POST /payments/sepa-ct/v1/customer-payments** endpoint so that the `Creditor` and `Debtor` fields no longer require a PostalAddress.  
If a PostalAddress is provided, then the `Country` and `TownName` must be provided.

The `Country` _and_ `TownName` fields were already mandatory, this change corrects the documentation for SCT and SCT Inst to align with the implementation.